Use set_enabled() and set_disabled() instead
When parsing the torrc file, we switch from command line flavor to using the python functions availabed in anon_connection_wizard package.

The advantages are as follows:

1. Better consistency with anon_connection_wizard style parsing

While the command line flavor parsing comments (out) `#DisableNetwork 0`, anon_connection_wizard writes `DisableNetwork 0` or `DisableNetwork 1` to 40_connection_wizard.conf . This can be a issue when user uses both anon-connection-wizard and whonixsetup for at least once. Using the same implementation will aviod the inconsistency.

2. More robust in handling different cases

Considering the inconsistency descriped above, the command line flavor parsing was not robust enough. For example, it will assume the final value in anon_connection_wizard.conf is 0 right after it sees a `DisableNetwork 0` line. This will ignore the case where there is also a `DisableNetwork 1` line after it.

3. Lower maintanance cost

It allows us to only maintance one piece of code.

4. Simplified the logic in the bash script

The disadvantages are as follows:

1. anon-connection-wizard becomes a dependency of whonixsetup now

However, since whonixsetup should be installed in both Whonox-Gateway and Whonox-Workstation, but anon-connection-wizard should NOT be installed in Whonix-Worksatation, we cannot claim the dependency in Debian control.

2. Less error info in CLI

The error happens in the python function may not be well-reflected to the CLI which is seen by user. For example, user may only know the attempt to enable Tor failed, but does not know if the problem is because of the missing of anon_connection_wizard package or anything else.